+++ internal/framework/provider/provider.go
@@ -20,6 +20,7 @@ import (
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/internal/framework/services/compute"
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/internal/framework/services/inmemorydbv2"
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/internal/framework/services/kafka"
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/internal/framework/services/mariadbv2"
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/internal/framework/services/monitoring"
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/internal/framework/services/objectstorage"
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/internal/framework/services/objectstoragemanagement"
@@ -277,6 +278,7 @@ func (p *IonosCloudProvider) Resources(_ context.Context) []func() resource.Reso
 		pgsqlv2.Resources(),
 		userobjectstorage.Resources(),
 		inmemorydbv2.Resources(),
+		mariadbv2.Resources(),
 	}
 
 	for _, r := range resources {
@@ -298,6 +300,7 @@ func (p *IonosCloudProvider) DataSources(_ context.Context) []func() datasource.
 		pgsqlv2.DataSources(),
 		userobjectstorage.DataSources(),
 		inmemorydbv2.DataSources(),
+		mariadbv2.DataSources(),
 	}
 
 	for _, r := range dataSources {
@@ -328,6 +331,7 @@ func (p *IonosCloudProvider) ListResources(_ context.Context) []func() list.List
 		objectstoragemanagement.ListResources(),
 		pgsqlv2.ListResources(),
 		inmemorydbv2.ListResources(),
+		mariadbv2.ListResources(),
 	}
 
 	for _, r := range listResources {